Many code tidying changes.  I'll try and do this on a file-by-file basis, so that any bugs I might have introduced are found.
Removed use of string module (it's deprecated, and its functionality is
completely supported by builtin string methods).

Added a new (and ugly, by comparison to the previous way, which I commented
out) way of writing out the pyscript font files.  This is to make sure that
the output is nicer than that generated by the previous technique.  This
only affects the write() method of the AFM class.

Removed tab indents.