Many code tidying changes.  I'll try and do this on a file-by-file basis, so that any bugs I might have introduced are found.
Removed use of string module.  Changed 'dict' variable to be called 'options'
(as that is what it refers to, and dict is a python internal word.  Fixed
calls to raise; they need an error type as the first argument.  Changed
(possibly incorrectly) some of the calls to the __init__ method of the
superclass to no longer use the apply() method.