OK, this will break some things .....
Firstly renamed Arrowhead to ArrowHead to be consistent with naming conventions in the
rest of the package.

Secondly, moved ArrowHead stuff into separate file arrowheads.py to avoid a recursive
import ... this didn't work so I've remoced again setting the default arrowhead in
defaults.

Thirdly, 'cleaned' up DoubleArrow but it not reversing the arrow at the start for some reason ....
will fix things later, go to go home ....