Added an anchor argument to the add_logo() method of the Talk class
... such that one can set the logo to appear at the top left or top
right of the page.  This *really* could do with improving in terms of
flexibility, but it works at present.  Had to stop adding an Area()
object to the logo when there is only one logo (why did I do this in the
first place????).  Added some error checking in the add_logo() method of
the Talk class and the _make() method of the Slide class to check for a
properly set anchor value.  (is this a good name???).  Changed the order
that the logos are added to the slide, such that now they should appear
over the signature and footer.