Initial import of optics components library to repository.
These are the individual eps files of the optics components library of
Alexander Franzen of the AEI in Hannover, Germany.  They were converted
to eps from his Adobe Illustrator component library and then passed
through eps2eps to reduce the file size (had a reduction factor of 221
for one file!).