Added a hacked (badly) add_titlepage_epsf() method to the Talk() class
... so that one can put images on title pages.  I sooooo need to go back
and redesign the way talks are written, it's not that easy, and I'm the
developer!  Added and add_object() method to the Slide() class, this is
basically the same as the add_fig() method, but without the implicity
white rectangle behind.  Added a _make_objects() method to handle
creation and collection of objects within a Slide() object.